python/requirements.txt
# AM-D02 MuJoCo 仿真依赖
mujoco>=3.0.0
numpy>=1.20.0
rerun-sdk>=0.18.0
pyserial
